Unsqueeze for loss.

+++ knowledgeablestories/models/knowledgeable_story_model.py
@@ -472,7 +472,7 @@ class KnowledgeableStoriesModel(Model):
                         [torch.unsqueeze(torch.dot(context_state[0], t), dim=0) for t in final_state])
                     print(logit_scores)
 
-                    target_classes = torch.zeros_like(logit_scores)
+                    target_classes = torch.zeros_like(logit_scores).long()
                     target_classes[0] = 1  # First one is always the real sentence.
 
                     logits_log_softmax = self._log_softmax(logit_scores)
